Added design patterns

+++ src/main/java/com/example/application/repository/AppointmentRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AppointmentRepository extends JpaRepository<Appointment, BigInteger> {
+public interface AppointmentRepository extends JpaRepository<Appointment, Long> {
     List<Appointment> findAll();
 
 }
